# [feature-wip](multi-catalog) reuse hdfsFS and decode values in batch without breaking up by null