--- a/be/src/io/hdfs_file_reader.h
+++ b/be/src/io/hdfs_file_reader.h
@@ -17,11 +17,93 @@
 
 #pragma once
 
+#include <chrono>
+
 #include "gen_cpp/PlanNodes_types.h"
+#include "gutil/hash/hash.h"
 #include "io/file_reader.h"
 #include "io/hdfs_builder.h"
 
 namespace doris {
+
+class HdfsFsHandle {
+private:
+    // the number of referenced client
+    std::atomic<int> _ref_cnt;
+    // HdfsFsCache try to remove the oldest handler when the cache is full
+    std::atomic<uint64_t> _last_access_time;
+    // Client will set invalid if error thrown, and HdfsFsCache will not reuse this handler
+    std::atomic<bool> _invalid;
+
+    uint64_t _now() {
+        return std::chrono::duration_cast<std::chrono::milliseconds>(
+                       std::chrono::system_clock::now().time_since_epoch())
+                .count();
+    }
+
+public:
+    hdfsFS hdfs_fs;
+    // When cache is full, and all handlers are in use, HdfsFsCache will return an uncached handler.
+    // Client should delete the handler in such case.
+    const bool from_cache;
+
+    HdfsFsHandle(hdfsFS fs, bool cached)
+            : _ref_cnt(0), _last_access_time(0), _invalid(false), hdfs_fs(fs), from_cache(cached) {}
+
+    ~HdfsFsHandle() {
+        DCHECK(_ref_cnt == 0);
+        if (hdfs_fs != nullptr) {
+            // Even if there is an error, the resources associated with the hdfsFS will be freed.
+            hdfsDisconnect(hdfs_fs);
+        }
+        hdfs_fs = nullptr;
+    }
+
+    int64_t last_access_time() { return _last_access_time; }
+
+    void inc_ref() {
+        _ref_cnt++;
+        _last_access_time = _now();
+    }
+
+    void dec_ref() {
+        _ref_cnt--;
+        _last_access_time = _now();
+    }
+
+    int ref_cnt() { return _ref_cnt; }
+
+    bool invalid() { return _invalid; }
+
+    void set_invalid() { _invalid = true; }
+};
+
+// Cache for HDFS file system
+class HdfsFsCache {
+public:
+    static int MAX_CACHE_HANDLE;
+
+    static HdfsFsCache* instance() {
+        static HdfsFsCache s_instance;
+        return &s_instance;
+    }
+
+    // This function is thread-safe
+    Status get_connection(THdfsParams& hdfs_params, HdfsFsHandle** fs_handle);
+
+private:
+    std::mutex _lock;
+    std::unordered_map<uint64, std::unique_ptr<HdfsFsHandle>> _cache;
+
+    HdfsFsCache() = default;
+    HdfsFsCache(const HdfsFsCache&) = delete;
+    const HdfsFsCache& operator=(const HdfsFsCache&) = delete;
+
+    uint64 _hdfs_hash_code(THdfsParams& hdfs_params);
+    Status _create_fs(THdfsParams& hdfs_params, hdfsFS* fs);
+    void _clean_invalid();
+    void _clean_oldest();
+};
 
 class HdfsFileReader : public FileReader {
 public:
@@ -47,7 +129,6 @@ public:
     virtual bool closed() override;
 
 private:
-    Status connect();
     void _parse_properties(const std::map<std::string, std::string>& prop);
 
 private:
@@ -58,7 +139,9 @@ private:
     int64_t _file_size;
     hdfsFS _hdfs_fs;
     hdfsFile _hdfs_file;
-    HDFSCommonBuilder _builder;
+    HdfsFsHandle* _fs_handle = nullptr;
+    bool _closed = false;
+    bool _opened = false;
 };
 
 } // namespace doris