… be/src/vec/exec/format/parquet/parquet_common.cpp
… be/src/vec/exec/format/parquet/parquet_common.h
… be/test/vec/exec/parquet/parquet_thrift_test.cpp