# Proposed changes

PR(https://github.com/apache/doris/pull/13404) introduced that ParquetReader will break up batch insertion when encountering null values, which leads to the bad performance compared to OrcReader. So this PR has pushed null map into decode function, reduce the time of virtual function call when encountering null values.

Further more, reuse hdfsFS among file readers to reduce the time of building connection to hdfs.

## Performance Test
Using one thread to query data in table with 10 million rows, and the table is generated from a [data generator tool](https://github.com/glebkorolkov/datagen) with all types supported by hive, and all values are random, and has a 10% probability of being null.
<meta charset="utf-8"><div data-page-id="N3sQdKUlGoZo3Vx3jdOcXp2UnYc" data-docx-has-block-data="true"><div>

| Type/Time(s) | Before Opt. | After Opt. |
| -- | -- | -- |
| tinyint | 0.34 | 0.14 |
| smallint | 0.34 | 0.14 |
| int | 0.32 | 0.13 |
| bigint | 0.34 | 0.15 |
| boolean | 0.31 | 0.13 |
| float | 0.32 | 0.12 |
| double | 0.32 | 0.14 |
| string | 2.41 | 2.31 |
| binary | 0.86 | 0.63 |
| timestamp | 1.12 | 0.88 |
| decimal | 0.40 | 0.20 |
| char | 0.44 | 0.21 |
| varchar | 0.44 | 0.21 |
| date | 1.12 | 0.91 |
| array<double> | 0.48 | 0.46 |
| array<string> | 2.33 | 1.86 |

## Checklist(Required)

1. Does it affect the original behavior: 
    - [ ] Yes
    - [x] No
    - [ ] I don't know
2. Has unit tests been added:
    - [ ] Yes
    - [x] No
    - [ ] No Need
3. Has document been added or modified:
    - [ ] Yes
    - [x] No
    - [ ] No Need
4. Does it need to update dependencies:
    - [ ] Yes
    - [x] No
5. Are there any changes that cannot be rolled back:
    - [ ] Yes (If Yes, please explain WHY)
    - [x] No

## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...