@hello-stephen 2022-10-27
TeamCity pipeline, clickbench performance test result:
 the sum of best hot time: 38.56 seconds
 load time: 555 seconds
 storage size: 17154644123 Bytes
 https://doris-community-test-1308700295.cos.ap-hongkong.myqcloud.com/tmp/20221027023643_clickbench_pr_34509.html