@hello-stephen 2022-10-27
TeamCity pipeline, clickbench performance test result:
 the sum of best hot time: 39.53 seconds
 load time: 575 seconds
 storage size: 17154644830 Bytes
 https://doris-community-test-1308700295.cos.ap-hongkong.myqcloud.com/tmp/20221028012100_clickbench_pr_34948.html